nira/requirements.txt
requests>=2.32.0
psutil>=5.9.8
pywin32>=306
cryptography>=42.0.0
pillow>=10.3.0
pytesseract>=0.3.10
fastapi>=0.115.0
uvicorn>=0.30.0
mysql-connector-python>=9.0.0
